Add an abstract delegate class that can be subclassed to provide a custom processing of search requests.
Up until now, if the client had wanted to perform a search request, it
would have had to grab URL, query parameters and request method and set
it up all on its own.
Now the OpenSearchEngine class will do the job, leaving only the final
part up to the client (i.e. how a request should be performed).

One can still use it in the old way, of course.